Move tests under composers subpackage and reuse golden files
