+++ src/test/java/com/google/api/generator/spring/composer/SpringPropertiesClassComposerTest.java
@@ -12,36 +12,35 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.api.generator.spring;
+package com.google.api.generator.spring.composer;
 
-import com.google.api.generator.engine.ast.ClassDefinition;
 import com.google.api.generator.gapic.composer.common.TestProtoLoader;
 import com.google.api.generator.gapic.model.GapicClass;
 import com.google.api.generator.gapic.model.GapicContext;
-import com.google.api.generator.spring.composer.SpringComposer;
+import com.google.api.generator.gapic.model.Service;
 import com.google.api.generator.test.framework.Assert;
-import java.util.List;
+import java.util.Arrays;
 import org.junit.Before;
 import org.junit.Test;
 
-public class SpringComposerTest {
+public class SpringPropertiesClassComposerTest {
   private GapicContext context;
+  private Service echoProtoService;
 
   @Before
   public void setUp() {
     this.context = TestProtoLoader.instance().parseShowcaseEcho();
+    this.echoProtoService = this.context.services().get(0);
   }
 
   @Test
-  public void spring_composer_test() {
-
-    List<GapicClass> gapicClasses = SpringComposer.composeServiceAutoConfigClasses(context);
-
-    // write to verify result for now
-    for (GapicClass gapicClazz : gapicClasses) {
-      ClassDefinition clazz = gapicClazz.classDefinition();
-      String fileName = clazz.classIdentifier() + ".golden";
-      Assert.assertGoldenClass(this.getClass(), gapicClazz, fileName);
-    }
+  public void generateAutoConfigClazzTest() {
+    GapicClass clazz =
+        SpringPropertiesClassComposer.instance().generate(this.context, this.echoProtoService);
+    GapicClass clazzWithHeader = SpringComposer.addApacheLicense(Arrays.asList(clazz)).get(0);
+    Assert.assertGoldenClass(
+        this.getClass(),
+        clazzWithHeader,
+        clazzWithHeader.classDefinition().classIdentifier() + ".golden");
   }
 }
