+++ src/test/java/com/google/api/generator/spring/composer/goldens/EchoSpringAutoConfiguration.golden
